backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
openai==1.3.7
whisper-openai==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
django-rest-knox==4.2.0